+++ CHANGELOG.md
@@ -17,6 +17,7 @@ Given a version number MAJOR.MINOR.PATCH, increment:
 - Ledger resource
 - LedgerTransaction resource
 - Parse method to PixPullSubscription
+- BusinessAccountRequest resource
 
 ## [0.13.0] - 2026-06-29
 ### Added